fix: 알림 설정 모달로 이전
