+++ src/pages/mypage/modals/NotificationSettingsModal.vue
@@ -1,11 +1,11 @@
 <template>
   <div class="notificationModalOverlay">
     <div class="notificationModal">
-      <div class="notificationSettings">
-        <!-- 헤더 -->
+      <div class="notification-settings">
+        <!-- 💪(상일) 헤더 -->
         <div class="header">
           <div class="headerSpacer"></div>
-          <div class="title font-18 font-bold">알림 설정</div>
+          <h2 class="title font-18 font-bold">알림 설정</h2>
           <img
             src="@/assets/images/icons/common/x.png"
             alt="닫기"
@@ -13,88 +13,91 @@
             @click="goBack"
           />
         </div>
-        <!-- 알림 권한 안내 -->
-        <div v-if="showPermissionNotice" class="permissionNotice">
+
+        <!-- 💪(상일) 알림 권한 안내 -->
+        <div v-if="showPermissionNotice" class="permission-notice">
           <p class="font-13">{{ permissionMessage }}</p>
         </div>
 
-        <!-- 알림 설정 리스트 -->
-        <div class="settingsList">
-          <div class="settingItem">
-            <div class="settingInfo">
-              <div class="settingTitle font-15 font-bold">북마크 정책 알림</div>
-              <p class="settingDesc font-12">
+        <!-- 💪(상일) 알림 설정 리스트 -->
+        <div class="settings-list" :class="{ 'loading-overlay': isTokenGenerating }">
+          <!-- 💪(상일) 로딩 스피너 -->
+          <div v-if="isTokenGenerating" class="settings-loading-spinner"></div>
+          <div class="setting-item">
+            <div class="setting-info">
+              <div class="setting-title font-15 font-bold">북마크 정책 알림</div>
+              <p class="setting-desc font-12">
                 북마크한 정책의 신청일 및 마감일을 알려드려요
               </p>
             </div>
             <button
-              class="toggleBtn font-11 font-bold"
+              class="toggle-btn font-11 font-bold"
               :class="{
                 on: subscriptionStatus.isActiveBookmark,
                 off: !subscriptionStatus.isActiveBookmark,
               }"
               @click="toggleNotification('bookmark')"
-              :disabled="!hasNotificationPermission || loading"
+              :disabled="!hasNotificationPermission || loading || isTokenGenerating"
             >
               {{ subscriptionStatus.isActiveBookmark ? 'ON' : 'OFF' }}
             </button>
           </div>
 
-          <div class="settingItem">
-            <div class="settingInfo">
-              <div class="settingTitle font-15 font-bold">TOP3 추천 알림</div>
-              <p class="settingDesc font-12">
+          <div class="setting-item">
+            <div class="setting-info">
+              <div class="setting-title font-15 font-bold">TOP3 추천 알림</div>
+              <p class="setting-desc font-12">
                 머니버니가 추천하는 맞춤 정책 TOP3를 알려드려요
               </p>
             </div>
             <button
-              class="toggleBtn font-12 font-bold"
+              class="toggle-btn font-11 font-bold"
               :class="{
                 on: subscriptionStatus.isActiveTop3,
                 off: !subscriptionStatus.isActiveTop3,
               }"
               @click="toggleNotification('top3')"
-              :disabled="!hasNotificationPermission || loading"
+              :disabled="!hasNotificationPermission || loading || isTokenGenerating"
             >
               {{ subscriptionStatus.isActiveTop3 ? 'ON' : 'OFF' }}
             </button>
           </div>
 
-          <div class="settingItem">
-            <div class="settingInfo">
-              <div class="settingTitle font-15 font-bold">신규 정책 알림</div>
-              <p class="settingDesc font-12">
+          <div class="setting-item">
+            <div class="setting-info">
+              <div class="setting-title font-15 font-bold">신규 정책 알림</div>
+              <p class="setting-desc font-12">
                 조건에 맞는 신규 정책이 등록되면 알려드려요
               </p>
             </div>
             <button
-              class="toggleBtn font-12 font-bold"
+              class="toggle-btn font-11 font-bold"
               :class="{
                 on: subscriptionStatus.isActiveNewPolicy,
                 off: !subscriptionStatus.isActiveNewPolicy,
               }"
               @click="toggleNotification('newPolicy')"
-              :disabled="!hasNotificationPermission || loading"
+              :disabled="!hasNotificationPermission || loading || isTokenGenerating"
             >
               {{ subscriptionStatus.isActiveNewPolicy ? 'ON' : 'OFF' }}
             </button>
           </div>
 
-          <div class="settingItem">
-            <div class="settingInfo">
-              <div class="settingTitle font-15 font-bold">소비 피드백 알림</div>
-              <p class="settingDesc font-12">
+          <div class="setting-item">
+            <div class="setting-info">
+              <div class="setting-title font-15 font-bold">소비 피드백 알림</div>
+              <p class="setting-desc font-12">
                 주간 소비 패턴 분석 리포트를 알려드려요
               </p>
             </div>
             <button
-              class="toggleBtn font-12 font-bold"
+              class="toggle-btn font-11 font-bold"
               :class="{
                 on: subscriptionStatus.isActiveFeedback,
                 off: !subscriptionStatus.isActiveFeedback,
               }"
               @click="toggleNotification('feedback')"
-              :disabled="!hasNotificationPermission || loading"
+              :disabled="!hasNotificationPermission || loading || isTokenGenerating"
             >
               {{ subscriptionStatus.isActiveFeedback ? 'ON' : 'OFF' }}
             </button>
@@ -106,18 +109,17 @@
 </template>
 
 <script setup>
-import { ref, onMounted, defineEmits } from 'vue';
-import { useRouter } from 'vue-router';
-import { storeToRefs } from 'pinia';
-import { useNotificationStore } from '@/stores/notification';
-import { subscribeToPush } from '@/firebase/notificationPermission';
+import { ref, onMounted, computed, defineEmits } from "vue";
+import { storeToRefs } from "pinia";
+import { useNotificationStore } from "@/stores/notification";
+import { fcmTokenManager, TOKEN_STATES } from "@/firebase/FCMTokenManager";
 
-const router = useRouter();
 const emit = defineEmits(['close']);
 
+// 💪(상일) Pinia 스토어 사용 - reactive 객체는 직접 사용
 const notificationStore = useNotificationStore();
-const subscriptionStatus = notificationStore.subscriptionStatus;
-const { loading } = storeToRefs(notificationStore);
+const subscriptionStatus = notificationStore.subscriptionStatus; // reactive 객체 직접 사용
+const { loading } = storeToRefs(notificationStore); // loading만 ref로 사용
 const {
   fetchSubscriptionStatus,
   toggleNotificationType,
@@ -127,48 +129,95 @@ const {
 
 const hasNotificationPermission = ref(false);
 const showPermissionNotice = ref(false);
-const permissionMessage = ref('');
+const permissionMessage = ref("");
+const isTokenGenerating = ref(false); // 💪(상일) 토큰 발급 중 상태
 
+// 💪(상일) 모달 닫기
 const goBack = () => {
-  emit('close'); // 모달 닫기
+  emit('close');
 };
 
+// 💪(상일) 알림 권한 확인 - FCMTokenManager 사용으로 간소화
 const checkNotificationPermission = async () => {
   if (!('Notification' in window)) {
     showPermissionNotice.value = true;
     permissionMessage.value = '이 브라우저는 알림을 지원하지 않습니다.';
     return;
   }
-  const permission = Notification.permission;
-  hasNotificationPermission.value = permission === 'granted';
-  if (permission === 'default') {
-    showPermissionNotice.value = false;
-  } else if (permission === 'denied') {
-    showPermissionNotice.value = true;
-    permissionMessage.value = '브라우저 설정에서 알림 권한을 허용해주세요.';
-  } else {
-    showPermissionNotice.value = false;
-    const token = localStorage.getItem('fcm_token');
-    if (!token) {
-      try {
-        await subscribeToPush();
-        await createInitialSubscription();
-      } catch (error) {
-        console.error('FCM 토큰 발급 실패:', error);
-      }
+
+  try {
+    loading.value = true;
+    const tokenState = fcmTokenManager.getTokenState();
+
+    switch (tokenState) {
+      case TOKEN_STATES.ACTIVE:
+        hasNotificationPermission.value = true;
+        showPermissionNotice.value = false;
+        break;
+
+      case TOKEN_STATES.NEED_PERMISSION:
+        hasNotificationPermission.value = false;
+        if (Notification.permission === "default") {
+          // default 상태: 자동으로 권한 요청 (메시지 없음)
+          showPermissionNotice.value = false;
+          try {
+            await requestPermission();
+          } catch (error) {
+            console.log("사용자가 권한 거부:", error.message);
+            // 거부 후에는 안내 메시지 표시
+            showPermissionNotice.value = true;
+            permissionMessage.value = "기기의 알림 권한을 허용해주세요.";
+          }
+        } else {
+          // denied 상태: 안내 메시지 표시
+          showPermissionNotice.value = true;
+          permissionMessage.value = "기기의 알림 권한을 허용해주세요.";
+        }
+        break;
+
+      case TOKEN_STATES.NEED_TOKEN:
+        hasNotificationPermission.value = true;
+        showPermissionNotice.value = false;
+        // 토큰 자동 발급 및 초기 구독
+        try {
+          isTokenGenerating.value = true; // 로딩 시작
+          await fcmTokenManager.getValidToken();
+          await createInitialSubscription();
+          console.log("✅ 토큰 발급 및 초기 구독 완룜");
+        } catch (error) {
+          console.error("토큰 발급 실패:", error);
+          showPermissionNotice.value = true;
+          permissionMessage.value = "알림 설정 초기화에 실패했습니다.";
+        } finally {
+          isTokenGenerating.value = false; // 로딩 종료
+        }
+        break;
+
+      default:
+        hasNotificationPermission.value = false;
+        showPermissionNotice.value = false;
     }
+  } catch (error) {
+    console.error("권한 확인 실패:", error);
+    hasNotificationPermission.value = false;
+  } finally {
+    loading.value = false;
   }
 };
 
+// 💪(상일) 알림 권한 요청 및 초기 구독 설정 - 간소화
 const requestPermission = async () => {
   try {
-    await subscribeToPush();
+    loading.value = true;
+    isTokenGenerating.value = true; // 로딩 시작
+
+    // FCM 토큰 발급 (권한 요청 포함)
+    const token = await fcmTokenManager.getValidToken();
+
     hasNotificationPermission.value = true;
     showPermissionNotice.value = false;
-    const token = localStorage.getItem('fcm_token');
-    if (!token) {
-      throw new Error('FCM 토큰 발급 실패');
-    }
+
+    // 초기 구독 설정 (모든 알림 false로 시작)
     const initialSubscription = {
       token,
       isActiveBookmark: false,
@@ -176,20 +225,50 @@ const requestPermission = async () => {
       isActiveNewPolicy: false,
       isActiveFeedback: false,
     };
-    await notificationStore.updateSubscription(initialSubscription);
+
+    await updateSubscription(initialSubscription);
+
+    // 구독 상태 재조회
     await fetchSubscriptionStatus();
   } catch (error) {
-    console.error('알림 권한 요청 실패:', error);
-    await checkNotificationPermission();
+    console.error("알림 권한 요청 실패:", error);
+    showPermissionNotice.value = true;
+    permissionMessage.value = "알림 권한 요청에 실패했습니다.";
+  } finally {
+    loading.value = false;
+    isTokenGenerating.value = false; // 로딩 종료
   }
 };
 
+// 💪(상일) 알림 타입별 토글
 const toggleNotification = async (type) => {
   if (!hasNotificationPermission.value) {
-    alert('먼저 알림 권한을 허용해주세요.');
+    // 권한 요청 시도
+    try {
+      await requestPermission();
+      // 권한 허용 성공 시 해당 알림 설정 계속 진행
+      if (!hasNotificationPermission.value) {
+        return; // 여전히 권한 없으면 중단
+      }
+    } catch (error) {
+      console.warn("권한 요청 실패:", error);
+      return;
+    }
+  }
+
+  // 💪(상일) FCM 토큰 확인 (FCMTokenManager 사용)
+  try {
+    await fcmTokenManager.getValidToken(); // 토큰이 없으면 자동 발급
+  } catch (error) {
+    console.error("FCM 토큰 획득 실패:", error);
+    alert("알림 설정을 불러오는 중입니다. 잠시 후 다시 시도해주세요.");
     return;
   }
+
   try {
+    loading.value = true;
+
+    // 💪(상일) reactive 객체는 .value 없이 접근
     let currentStatus = false;
     switch (type) {
       case 'bookmark':
@@ -205,29 +284,66 @@ const toggleNotification = async (type) => {
         currentStatus = subscriptionStatus.isActiveFeedback;
         break;
     }
+
     await toggleNotificationType(type, !currentStatus);
+    console.log(`✅ ${type} 알림 설정 변경 완료: ${!currentStatus}`);
   } catch (error) {
-    console.error('알림 설정 변경 실패:', error);
-    alert('알림 설정 변경에 실패했습니다. 다시 시도해주세요.');
+    console.error("알림 설정 변경 실패:", error);
+    alert("알림 설정 변경에 실패했습니다. 다시 시도해주세요.");
+  } finally {
+    loading.value = false;
   }
 };
 
+// 💪(상일) 권한 변경 감지 및 자동 새로고침
+const setupPermissionWatcher = () => {
+  let lastPermission = Notification.permission;
+  
+  const checkPermissionChange = () => {
+    if (Notification.permission !== lastPermission) {
+      console.log(`🔄 알림 권한 변경 감지: ${lastPermission} → ${Notification.permission}`);
+      
+      // 권한 변경 시 즉시 새로고침
+      // granted → denied: 권한 해제
+      // denied → granted: 권한 허용
+      // default → granted/denied: 최초 권한 설정
+      window.location.reload();
+    }
+  };
+  
+  // 페이지 포커스 시 권한 상태 체크
+  document.addEventListener('visibilitychange', () => {
+    if (!document.hidden) {
+      setTimeout(checkPermissionChange, 100); // 잠시 대기 후 체크
+    }
+  });
+  
+  // 윈도우 포커스 시에도 체크
+  window.addEventListener('focus', () => {
+    setTimeout(checkPermissionChange, 100);
+  });
+};
+
+// 💪(상일) 컴포넌트 마운트 시 초기화
 onMounted(async () => {
   await checkNotificationPermission();
-  if (Notification.permission === 'default') {
-    try {
-      await requestPermission();
-    } catch (error) {
-      // 실패해도 페이지는 정상 로드
-    }
-  }
+  
+  // 💪(상일) 권한 변경 감지 설정
+  setupPermissionWatcher();
+
+  // 💪(상일) 알림 권한이 있을 때만 구독 상태 조회
   if (hasNotificationPermission.value) {
-    await fetchSubscriptionStatus();
+    try {
+      await fetchSubscriptionStatus();
+    } catch (error) {
+      console.warn("구독 상태 조회 실패:", error);
+    }
   }
 });
 </script>
 
 <style scoped>
+/* 💪(상일) 모달 컨테이너 (기존 유지) */
 .notificationModalOverlay {
   position: fixed;
   inset: 0;
@@ -250,12 +366,15 @@ onMounted(async () => {
   padding: 0;
   -webkit-overflow-scrolling: touch;
 }
-.notificationSettings {
+
+/* 💪(상일) NotificationSettings.vue 스타일 이전 */
+.notification-settings {
   background-color: #f8f9fa;
-  border-radius: 18px;
+  border-radius: 12px;
   min-height: unset !important;
 }
 
+/* 헤더 */
 .header {
   display: flex;
   align-items: center;
@@ -271,7 +390,9 @@ onMounted(async () => {
   flex: none;
   text-align: center;
   margin: 0 auto;
+  color: var(--text-login);
 }
+
 .closeIcon {
   width: 22px;
   height: 22px;
@@ -282,64 +403,124 @@ onMounted(async () => {
   cursor: pointer;
 }
 
-.permissionNotice {
+/* 권한 안내 */
+.permission-notice {
   margin: 20px;
   padding: 20px;
   background-color: #fff3cd;
-  border-radius: 10px;
+  border-radius: 12px;
   text-align: center;
 }
-.permissionNotice p {
+
+.permission-notice p {
   color: #856404;
   margin: 0;
 }
 
-.settingsList {
+.permission-btn {
+  margin-top: 12px;
+  padding: 10px 20px;
+  background-color: var(--text-green);
+  color: white;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+.permission-btn:hover {
+  background-color: #28a745;
+}
+
+.permission-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+/* 설정 리스트 - 연결된 스타일 및 로딩 처리 */
+.settings-list {
   padding: 20px;
+  position: relative;
   display: flex;
   flex-direction: column;
   gap: 0;
 }
 
-.settingItem {
-  display: grid;
-  grid-template-columns: 1fr auto;
+.loading-overlay {
+  opacity: 0.3;
+  pointer-events: none;
+}
+
+.settings-loading-spinner {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 40px;
+  height: 40px;
+  border: 3px solid #f3f3f3;
+  border-top: 3px solid var(--base-blue-dark);
+  border-radius: 50%;
+  animation: spin 1s linear infinite;
+  z-index: 10;
+  background-color: white;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
+}
+
+@keyframes spin {
+  0% { transform: translate(-50%, -50%) rotate(0deg); }
+  100% { transform: translate(-50%, -50%) rotate(360deg); }
+}
+
+.setting-item {
+  display: flex;
+  justify-content: space-between;
   align-items: center;
   padding: 13px 14px;
-  background-color: #fff;
+  background-color: white;
   border-radius: 12px;
   margin-bottom: 10px;
 }
 
-.settingInfo {
+.setting-info {
   flex: 1;
   margin-right: 5px;
 }
-.settingTitle {
+
+.setting-title {
   color: var(--text-login);
   margin: 5px;
 }
-.settingDesc {
+
+.setting-desc {
   color: var(--text-lightgray);
+  line-height: 1.4;
   margin: 5px;
 }
-.toggleBtn {
+
+/* 토글 버튼 */
+.toggle-btn {
   min-width: 52px;
   height: 32px;
   border-radius: 20px;
   border: none;
-  color: #fff;
-  background: var(--input-disabled-2);
+  color: white;
+  cursor: pointer;
+  transition: background-color 0.3s;
   letter-spacing: 0.04em;
   margin-left: 4px;
 }
-.toggleBtn.on {
+
+.toggle-btn.on {
   background-color: var(--text-green);
 }
-.toggleBtn.off {
+
+.toggle-btn.off {
   background-color: var(--text-lightgray);
 }
-.toggleBtn:disabled {
+
+.toggle-btn:disabled {
+  opacity: 0.5;
   cursor: not-allowed;
 }
 </style>
